Bugfix: called entry point twice for EXE load
The script used to call the entry point twice when doing an EXE load.
When loading a DLL, the entry is called twice (because the entry is
DllMain, and you have to tell DllMain when you attach and detach). For
EXEs, the entry point is the main function, so calling this twice will
result in the EXEs main function being called twice which is not the
intended behavior.